update some changes in event_category

+++ src/event_category/event_category.service.ts
@@ -2,6 +2,7 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { EventCategory } from './entities/event_category.entity';
+import { MESSAGE } from 'src/common/collection';
 
 @Injectable()
 export class EventCategoryService {
@@ -12,6 +13,13 @@ export class EventCategoryService {
 
   async createEventCategory(eventCategory: Partial<EventCategory>): Promise<EventCategory> {
     try {
+      const eventCategoryExist = await this.findEventByName(eventCategory.name);
+      if (eventCategoryExist) {
+        throw new HttpException({
+          status: HttpStatus.BAD_REQUEST,
+          error: MESSAGE.WARNING.EVENT_CATEGORY_EXIST
+        }, HttpStatus.BAD_REQUEST);
+      }
       const newCategory = this.eventCategoryRepository.create(eventCategory);
       return await this.eventCategoryRepository.save(newCategory);
     } catch (error) {
@@ -35,7 +43,14 @@ export class EventCategoryService {
 
   async findOneEventCategory(id: string): Promise<EventCategory> {
     try {
-      return await this.eventCategoryRepository.findOne({ where: { id } });
+      const eventCategory = await this.eventCategoryRepository.findOne({ where: { id } });
+      if (!eventCategory) {
+        throw new HttpException({
+          status: HttpStatus.NOT_FOUND,
+          error: MESSAGE.WARNING.EVENT_CATEGORY_NOT_FOUND
+        }, HttpStatus.NOT_FOUND);
+      }
+      return eventCategory;
     } catch (error) {
       throw new HttpException({
         status: error.status ? error.status : HttpStatus.INTERNAL_SERVER_ERROR,
@@ -46,6 +61,13 @@ export class EventCategoryService {
 
   async updateEventCategory(id: string, updateData: Partial<EventCategory>): Promise<EventCategory> {
     try {
+      const eventCategory = await this.findOneEventCategory(id);
+      if (!eventCategory) {
+        throw new HttpException({
+          status: HttpStatus.NOT_FOUND,
+          error: MESSAGE.WARNING.EVENT_CATEGORY_NOT_FOUND
+        }, HttpStatus.NOT_FOUND);
+      }
       await this.eventCategoryRepository.update(id, updateData);
       return await this.eventCategoryRepository.findOne({ where: { id } });
     } catch (error) {
@@ -58,6 +80,13 @@ export class EventCategoryService {
 
   async removeEventCategory(id: string): Promise<void> {
     try {
+      const eventCategory = await this.findOneEventCategory(id);
+      if (!eventCategory) {
+        throw new HttpException({
+          status: HttpStatus.NOT_FOUND,
+          error: MESSAGE.WARNING.EVENT_CATEGORY_NOT_FOUND
+        }, HttpStatus.NOT_FOUND);
+      }
       await this.eventCategoryRepository.delete(id);
     } catch (error) {
       throw new HttpException({
